requirements.txt
streamlit
ultralytics
pillow
pandas
plotly
requests